# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MaxLife Realty API",
    "description": "Public API for MaxLife Realty — Florida's statewide commercial real estate brokerage. Endpoints for investor tools (deal analysis, net sheets), lead capture, marketplace inquiries, and semantic listing search. Base URL: https://maxlifedevelopment.com",
    "version": "1.0.0",
    "contact": {
      "name": "MaxLife Realty",
      "email": "Ryan@MaxLifeRealty.com",
      "url": "https://maxlifedevelopment.com/contact"
    }
  },
  "servers": [
    {
      "url": "https://maxlifedevelopment.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/search/listings": {
      "post": {
        "tags": ["search"],
        "summary": "Semantic natural-language listing search",
        "description": "Accepts a natural-language query (e.g. 'passive NNN property near airport under $3M') and returns ranked marketplace listings using vector similarity search (pgvector). Requires OPENAI_API_KEY or VOYAGE_API_KEY to be configured on the server. Falls back to full-text search if embeddings are unavailable.",
        "operationId": "semanticSearchListings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SemanticSearchRequest" },
              "example": {
                "query": "passive NNN investment near airport under $3M with long lease term",
                "limit": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked listing results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SemanticSearchResponse" }
              }
            }
          },
          "400": { "description": "Missing or empty query", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "type": "string", "example": "Valid email required" } },
        "required": ["error"]
      },
...
      "SemanticSearchRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": { "type": "string", "minLength": 2, "description": "Natural-language description of what you're looking for", "example": "passive NNN investment near airport under $3M with long lease term" },
          "limit": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 }
        }
      },
      "ListingSummary": {
        "type": "object",
        "description": "Marketplace listing summary returned by search",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "property_type": { "type": "string" },
          "transaction_type": { "type": "string", "enum": ["sale", "lease", "both"] },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "price": { "type": "number", "nullable": true },
          "cap_rate": { "type": "number", "nullable": true },
          "building_sqft": { "type": "integer", "nullable": true },
          "tenant": { "type": "string", "nullable": true },
          "lease_type": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "under_contract", "sold"] },
          "similarity": { "type": "number", "description": "Cosine similarity score (0–1). Only present for vector search results.", "nullable": true },
          "url": { "type": "string", "description": "Full URL to the listing detail page", "example": "https://maxlifedevelopment.com/marketplace/dollar-general-kissimmee-fl" }
        }
      },
      "SemanticSearchResponse": {
        "type": "object",
        "properties": {
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/ListingSummary" } },
          "total": { "type": "integer" },
          "mode": { "type": "string", "enum": ["vector", "fulltext"], "description": "Whether vector similarity or full-text search was used" }
        }
      }
    }
  }
}